pyproject.toml
[tool.poetry]
name = "zoondb"
version = "2021.06-alpha.0"
description = "CHIME/FRB Backend - zoondb"
authors = ["Gordon Hamilton <gordon.hamilton1110@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.1.0"
chime-frb-api = "^2021.5.28"
motor = "^2.4.0"
maestro = {git = "git@github.com:chimefrb/maestro", rev = "main", extras = ["mongodb"]}
numpy = "^1.20.3"
python-dotenv = "^0.18.0"
tqdm = "^4.61.1"
panoptes-client = "^1.4.0"
nest-asyncio = "^1.5.1"
mongoengine = "^0.23.1"
libmagic = "^1.0"
python-magic-bin = "0.4.14"
schedule = "^1.1.0"
pytest = "^6.2.4"
pytest-sanic = "1.6.2"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
mkdocs-material = "^5.5.3"
pre-commit = "^2.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
